Main growth analysis for Taxon Profiles

Select measures, remove outliers, select the best growth model and produce the percentile of the best fitting growth model for the sexes and birth types selected.

Gro_Main(
  data,
  coresubse,
  Taxa,
  Species,
  BirthType = "All",
  AgeMat = NA,
  type = "weight",
  MeasureType = NULL,
  MinInstitution = 1,
  UncertDate = 365,
  MinNGro = 30,
  MinNIGro = 30,
  models = "vonBertalanffy",
  percentiles = c(2.5, 97.5),
  PlotDir = NULL,
  MinDate = "1980-01-01",
  PlotName = ""
)

Arguments

data

data.frame including the following columns MeasurementValue, MeasurementDate, UnitOfMeasure, EstimatedMeasurement, RecordType, MeasurementType, AnimalAnonID and RecordingInstitution.

coresubse

data.frame including at least the following columns binSpecies, SexType, AnimalAnonID, BirthType and Birthdate (date)

Taxa

character the name of the taxa studied

Species

character the name of the species studied

BirthType

vector of character "Captive", "Wild", and/or "All"

AgeMat

numeric age at sexual maturity.

type

character Either 'weight' or 'length'.ht

MeasureType

vector of characters Name of the type of measurements that should be included. Default: all measurement types are included.

MinInstitution

numeric Minimum number of institutions that should hold records to run the analysis.

UncertDate

numeric: Maximum uncertainty accepted for measurement dates, in days

MinNGro

numeric Minimum number of weight needed to fit the growth models

MinNIGro

numeric Minimum number of unique individuals needed to fit the growth models

models

vector of characters indicating the growth models that need to be fit.The following models are supported : logistic, gompertz, chapmanRichards, vonBertalanffy, polynomial.

percentiles

vector of numeric indicating the percentiles that need to be estimated. default = c(2.5,97.5) corresponding to the 95% predicted interval.

PlotDir

character Directory to save the plots. Default: no plot is saved.

MinDate

character 'YYYY-MM-DD' Earlier date to include data

PlotName

character Name used to save the plot.

Value

A list including an analysis by sex and by birth type as included in the input. Each analysis includes:

The fit of the growth models with the AIC table of the model, the fit of the best model, a data frame with the percentile selected and 3 GOF test including the normality of residuals and the test of trend in residuals against X and in the variance of the residuals.

A summary with the number of individuals and weight used for the analysis:

NInd_raw and, NWeight_raw indicate the number of unique individuals and the number of weights once the sex, the birth type, the measurment type from global collection have been selected

NWeight_val and NInd_val indicate the number of unique individuals and the number of weights once valid measurment from Husbandry have been selected

NJuv= 0 and NAd indicate the number of weights of juveniles and adults selected

NJuv_keep= 0 and NAd_keep indicate the number of weights of juveniles and adults selected after removing outliers

NWeight and NInd indicate the total number of unique individuals and the number of weights used for the analysis

a logical indicated if the growth analysis was performed

If the growth analysis was not performed, an error and its number (Nerr) are returned: The possibility for this functions are: 1/No raw data and 2/No valid weight measure 3/Data from 1 Institution 4/ NWeight < MinNGro #' 5/NInd < MinNIGro 6/Best model did not fit. If PlotDir is filled, 2 plots are produced: one showing the outliers removed from the data, and one showing the fit of the model on the data.

Examples

data(weights)
data(core)
output= Gro_Main(data = weights, coresubse = core,
                 Taxa = "Reptilia", Species = "Testudo hermanni" ,
                 BirthType = c("Captive", "Wild"), AgeMat = 1,
                 type = "weight", MeasureType = "Live weight",
                 MinNGro = 30, MinNIGro = 30, 
                 models = c("vonBertalanffy", "logistic"), percentiles = c(2.5,97.5)) 
#> Warning: Multiple measurment types found in the data: Live weight, Data should often include only one Measurement Type
#> Warning: Multiple measurment types found in the data: Live weight, Data should often include only one Measurement Type