Growth model selection

This function fits a series of growth models to a dataset, select the best one by AIC and estimates the percentiles of the predicted distribution of values.

Gro_analysis(
  data_weight,
  all_mods = c("vonBertalanffy"),
  percentiles = c(2.5, 97.5)
)

Arguments

data_weight

data.frame including at least the numeric columns Age, MeasurementValue and AnimalAnonID.

all_mods

vector of character indicatingthe growth models that need to be fit.The following models are supported : logistic, gompertz, chapmanRichards, vonBertalanffy, polynomial.

percentiles

vector of numeric indicating the percentiles that need to be estimated. default: 95% predicted interval.

Value

a list including:

a data frame with the percentile selected

the fit of the best model

the AIC table of the model

GOF: 3 GOF test including the normality of residuals and the test of trend in residuals against X and in the variance of the residuals

Examples

Age <- sample(c(0:10), 100, replace = TRUE)
AnimalAnonID <- sample(c(0:20), 100, replace = TRUE)
MeasurementValue <- exp(0.2+15 * (1 - exp(-(0.1) * log(Age+1)))+ 
                          rnorm(100,0,0.01) + AnimalAnonID*0.1)-1 
dat = data.frame(Age = Age, MeasurementValue = MeasurementValue, 
                 AnimalAnonID = AnimalAnonID, MeasurementType = "Live Weight")

out = Gro_analysis(dat, 
                   all_mods = c("logistic", "vonBertalanffy", "gam"), 
                   percentiles = c(2.5, 97.5))