Prepare weight and length Dataset

Select clean measures, unify the unity used and add individual age at each measure

Gro_cleanmeasures(
  data,
  coresubse,
  Birth_Type = "All",
  type = "weight",
  MeasureType = NULL,
  mindate = "1980-01-01",
  uncert_date = 365,
  corevariable = NULL,
  variablekeep = NULL
)

Arguments

data

data.frame including the following columns MeasurementValue, MeasurementDate, MeasurementDateStart, MeasurementDateEnd, UnitOfMeasure, EstimatedMeasurement, RecordType, MeasurementType, AnimalAnonID, RecordingInstitution, and Age.

coresubse

data.frame including at least the following columns binSpecies, SexType, AnimalAnonID and BirthType

Birth_Type

character Captive, Wild, or All Default = TRUE

type

character Either 'weight' or 'length'. Default = 'weight

MeasureType

vector of characters Name of the type of measurements that should be included. Default = NULL, all measurement type are included.

mindate

character 'YYYY-MM-DD' Earlier date to include data

uncert_date

numeric: Maximum uncertainty accepted for measurement dates, in days

corevariable

vector character Name of the core columns that must be added to data

variablekeep

vector character Name of the data columns that must also be returned

Value

A list including

The data frame including selected measures, plus individual birth date and individual age at each measure

A summary with the number of individuals and weight: - NInd_raw and, NWeight_raw indicate the number of unique individuals and the number of weights once the sex, the birth type, the measurment type from global collection have been selected - NWeight_val and NInd_val indicate the number of unique individuals and the number of weights once valid measurment from Husbandry have been selected - If no data are selected, an error and its number (Nerr) are returned: The possibility for this functions are: 1/No raw data and 2/No valid weight measure.

Examples

data(weights)
data(core)
data_weights= Gro_cleanmeasures(weights, core,
                                Birth_Type = "All",
                                MeasureType = 'Live weight',
                                mindate = "1980-01-01")