Remove outliers from weigth or length data

Takes a data frame including weight measures and Age and look for possible outliers

Gro_remoutliers(
  data_weight,
  taxa,
  ageMat = NULL,
  maxweight = NULL,
  variableid = "AnimalAnonID",
  min_Nmeasures = 7,
  perc_weight_min = 0.2,
  perc_weight_max = 2.5,
  IQR = 2.75,
  minq = 0.025,
  Ninterval_juv = 10
)

Arguments

data_weight

data.frame including at least the following columns : MeasurementValue (numeric), Age (numeric)

taxa

character the name of the taxa studied

ageMat

numeric the age at sexual maturity to differentiate juveniles (still growing) from adults

maxweight

numeric the maximum weight allowed for the dataset

variableid

character name of the variable including individual ids. Defaut = "AnimalAnonID". It must also be a column of data_weight.Default = 6. It must be at least 5

min_Nmeasures

integer Minimum number of measures for an individual to check outliers along its growth trajectory. Default = 7. It must be at least 5

perc_weight_min

numeric Minimum percentage of weight that an individual can naturally lose or gain during a year . Default = 0.2 (20%)

perc_weight_max

numeric Maximum percentage of weight that an individual can naturally lose or gain during a year . Default = 2.5 (250%)

IQR

numeric influences the sensitivity of the function to remove outliers using log-linear model of growth. It should be above 1 Default = 2.75. A higher number makes the function less sensitive to find outliers.

minq

numeric Sensitivity of the function to remove outliers using percentiles, between 0 and 1. Default = 0.025

Ninterval_juv

integer Number of intervals used for the sliding windows for juveniles. Default = 10

Value

The data frame including the additional column KEEP a numeric vector of 1 and 0 indicating the measures to keep. The 0 signal outliers. Other additional columns keep1, keep2, keep3 indicates the individuals highlighted as outliers (0) in steps 1 to 3 and the column juv indicates which individuals have been classified as juveniles.

Details

This function follow 4 different steps to highlight outliers: 1/ It removes the measures higher than maxweight (Instead of using the OrdMag Fernando that was removing only very extreme values, a maximum value was set for each taxa). 2/ Independently on adults and juveniles, it uses the function Gro_Rout_quan() to removes outliers based on percentiles on adults and juveniles. It uses sliding windows of age for juveniles. 3/ Independently on adults and juveniles, it uses the function Gro_lin_ind() to build generalized additive models for each individual trajectories with at least 7 measures and remove outliers based on the residuals of the models. 4/ It uses the function Gro_lin_ind() to build a common generalized additive model for all growth trajectories and remove outliers based on the residuals of the models.

Examples

data(weights)
weights = Gro_remoutliers(weights, taxa = "Mammalia", ageMat = 10)