Prepare Animal data

Check dates and add needed columns to animal data

Prep_Animal(Animal, minDate = "1900-01-01", extractDate)

Arguments

Animal

data.frame including at least the following columns AnimalAnonID, BirthDate, BirthDateEstimateType, BirthDateEstimateStart, BirthDateEstimateEnd, BirthType, FirstAcquisitionDate, DeathDate, DeathDateEstimateType, DeathDateEstimateStart, DeathDateEstimateEnd, LastCommentEntryDate, BirthObserved, GlobalStatus, and LastTXDate

minDate

character: Earlier possible date: date used when minimum birth or death date unknown

extractDate

character: Latest possible date in the data: date used when individuals are still alive

Value

The checked and prepared dataset

Details

This function add the columns: MinBirthDate, MaxBirthDate, MinDeathDate, MaxDeathDate, EntryDate, EntryType, DepartDate, DepartType, Birth_Uncertainty, and Death_Uncertainty. It removes individuals for which dates are not chronologically meaningful (i.e. birth date after death date) and for which Birth_Uncertainty is higher than uncert_birth.

Examples

file = system.file("sci_Animal.csv", package = 'ISRverse')
ZIMSdirtest = dirname(file)

data <- Load_Zimsdata  (taxa = "Reptilia",
                       ZIMSdir = ZIMSdirtest,
                       Animal = TRUE)
#> ReptiliaReptilia Done.

Animal <- Prep_Animal(data$Reptilia$Animal, extractDate = lubridate::as_date("2023/12/23"))

unlink(ZIMSdirtest, recursive = FALSE)