Litter Size

Make a data frame of litter size per parent and birth date and produce summary tables of litter size. it selects only offspring/mother couples

Rep_littersize(
  Reprodata,
  perAge = FALSE,
  Nday = 7,
  parentProb = 80,
  minNlitter = 30
)

Arguments

Reprodata

data frame including at least the columns AnimalAnonID, ParentAnonID, ParentType, Probability, Offspring_BirthDate, Offspring_Inst and Parent_Age

perAge

logical Whether to estimate litter size per mother age. Default = FALSE

Nday

numeric Number of consecutive days over which the birth dates of a litter/clutch can be spread. Default = 7

parentProb

numeric Minimum percentage of parentage probability to include. Default = 80

minNlitter

numeric Minimum number of litters to run the analysis. The data frame for litter size will be produced in all cases. Default = 30

Value

A list including:

A data frame with the litter size per mother and birth date

A table including the number and proportion of each size of litter

Summary statistics including:

NOffsp_prob: Number of offspring

NParent_prob: Number of parents

NReprEvent: Number of litters

MeanLittSize: Mean litter size

MedLittSize: Median litter size

SdLittSize: Standard deviation of litter size

analyzed = FALSE,

error: if the litter size where not analysze, the associated error is printed here. It can be :"No births with parentage probability higher than parentProb or N litter < minNlitter

Nerr: The number id of the error

if perAge = TRUE, a table including the mean and standard deviation of litter size per maternal age

Examples

data(core)
data(collection)
data(parent)
data(moves)
#prepare Data
Data <- Rep_prepdata (coresubset = core, collection, parent, moves, minNrep = 1, minNparep = 1,
                      BirthType_parent = "All", BirthType_offspring = "All" )

#Estimate litter size
out <- Rep_littersize(Data$Reprodata, perAge = FALSE,
                      Nday = 7)

out
#> $summary
#> $summary$NOffsp_prob
#> [1] 6
#> 
#> $summary$NParent_prob
#> [1] 4
#> 
#> $summary$NReprEvent
#> NULL
#> 
#> $summary$analyzed
#> [1] FALSE
#> 
#> $summary$error
#> N litter < 30
#> 
#> $summary$Nerr
#> [1] 5
#> 
#> $summary$MeanLittSize
#> NULL
#> 
#> $summary$MedLittSize
#> NULL
#> 
#> $summary$SdLittSize
#> NULL
#> 
#> 
#> $littSizeDf
#> # A tibble: 5 × 4
#>   MotherAnonID InstitutionAnonID litterSize MeanBirthDate
#>          <int>             <int>      <int> <date>       
#> 1      3752015            125738          1 1992-08-11   
#> 2      1378072             87126          1 2008-07-01   
#> 3      1378072             87126          2 2011-07-10   
#> 4      1552859            108112          1 1997-01-25   
#> 5      4224544            114122          1 2002-06-22   
#> 
#> $littSizeTab
#> NULL
#> 
#> $littSizeperAge
#> NULL
#>