Prepare Reproduction date

Prepare reproduction data including age of parent at birth, offspring sex and birth date and data of potential reproductive individuals at each age.

Rep_prepdata(
  coresubset,
  collection,
  parent,
  moves,
  BirthType_parent = "Captive",
  BirthType_offspring = "Captive",
  Age_uncert = 1e+06,
  Global = TRUE
)

Arguments

coresubset

data frame cleaned core data including only the selected individuals. It must includes at least the following columns: anonID, 'birthType', MaxBirthDate, MinBirthDate, DepartDate, BirthDate,Sex, and firstInst

collection

data frame Collection data including at least the following columns: AnimalID, ScopeType, ChangeDate

parent

data frame Parent data including at least the following columns: ParentAnonID, ParentCollectionScopeType, OffspringCollectionScopeType, AnonID, ParentOriginType, Probability

moves

data frame Moves data including at least the following columns: AnonID, To, Date

BirthType_parent

character Captive, Wild, or All. Default = "Captive"

BirthType_offspring

character Captive, Wild, or All. Default = "Captive"

Age_uncert

numeric Maximum uncertainty (= MaxBirthDate - MinBirthDate) in birth date in days. Default = 10000

Global

logical Whether only individuals belonging to global collections should be used. Default = TRUE

Value

A list including

The reproduction data

The potential reproductive Adults

A summary including:

Nbirths: the raw number of births

Nadults: the raw number of potential reproductive adults

NOffsp: the number of offspring with known birth date and with parents with known age

NParent: the number of unique parents with known age and with offspring with known birth date

NOffsp_age: the number of offspring with known birth date and with parents with a known age > 0

NParent_age: the number of unique parents with known and positive age and with offspring with known birth date

Examples

data(core)
data(collection)
data(parent)
data(moves)

Data <- Rep_prepdata (coresubset = core, collection, parent, moves,
                      BirthType_parent = "Captive", BirthType_offspring = "Captive"
)