First-year survival Analysis

Run basta models on the first year after checking conditions for correct convergence of the models

Sur_ana1(
  sexData,
  DeathInformation,
  models = "GO",
  shape = "simple",
  minNsur = 50,
  maxNsur = NULL,
  mindate = "1980-01-01",
  uncert_birth = 30,
  uncert_death = 30,
  minInstitution = 1,
  niter = 25000,
  burnin = 5001,
  thinning = 20,
  nchain = 3,
  ncpus = 2
)

Arguments

sexData

data.frame including at least the following columns AnimalAnonID, BirthDate, DepartDate, EntryDate, MaxBirthDate, MinBirthDate, EntryType, DepartType, FirstHoldingInstitution, LastHoldingInstitution, SexType, and BirthType

DeathInformation

data.frame including at least the following columns AnimalAnonID and RelevantDeathInformationType

models

vector of characters names of the basta models to run: "G0", "EX", "LO" and/or "WE". see ?basta for more information. Default = "GO"

shape

character shape of the basta model: "simple", "Makeham", "bathtub". see ?basta for more information. Default = "simple"

minNsur

numeric Minimum number of individual records needed to run the survival analysis. Default = 50

maxNsur

numeric Maximum number of individual records to run the survival analysis. Default = NULL

mindate

character 'YYYY-MM-DD' Earlier date to include data

uncert_birth

numeric: Maximum uncertainty accepted for birth date, in days

uncert_death

numeric: Maximum uncertainty accepted for death date, in days

minInstitution

numeric Minimum number of institutions that should hold records to run the survival analysis. Default = 1

niter

numeric. number of MCMC iterations to run. see ?basta for more information. Default = 25000

burnin

numeric Number of iterations removed so that the model has time to converge. see ?basta for more information. Default = 5001

thinning

numeric Number of iteration to run before saving a set of parameters. see ?basta for more information. Default = 20

nchain

numeric Number of chains to run. Default = 5001

ncpus

numeric Number of computer core to use. Default = 2

Value

The output of a list including:

a summary of the data used:

NGlobal: Number of captive born individuals selected from global collections

NBasta: Number of data (individuals) selected for the BaSTA/survival analysis

Ndead:Number of individuals with known age of death used in the BaSTA/survival analysis

maxAge: Maximum age of known age individuals

maxAlive:Maximum number of years spent ex situ

lxMin:Minimum survivorship reached with the raw Kaplan-Meier model

OutLev: threshold selected for the distribution of time spent alive: 100%, 99.9, 99 or 95%

a logical indicated if the growth analysis was performed

If the survival analysis was not performed, an error and its number (Nerr) are returned: The possibility for this functions are: 1/Nglobal < minNsur; 2/lxMin > minlx; 3/NBasta = 0; 4/ %known births < MinBirthKnown; 5/Data from 1 Institution; 6/Nbasta < minNsur; 7/Nbasta < maxNsur; 8/no DIC from Basta.

the basta fit of the best model

the DIC table comparing the different fit of the models

Examples

data(core)
data(deathinformation)
out <- Sur_ana1(core,  DeathInformation = deathinformation, models = "GO", shape = "simple",
                niter = 1000, burnin = 101, thinning = 10, nchain = 3, ncpus = 3)
#> [1] "GO"
#> 
#> Running sequence to find jump SDs... Done
#> Total jump SDs computing time: 10.97 secs.
#> 
#> Multiple simulations started...
#> 
#> snowfall 1.84-6.3 initialized (using snow 0.4-4): parallel execution on 3 CPUs.
#> Library BaSTA loaded.
#> Library BaSTA loaded in cluster.
#> 
#> Stopping cluster
#> Simulations finished.
#> Total MCMC computing time: 8.27 secs.
#> 
#> Calculating summary statistics... Done.
#> All parameters converged properly.
#> Constructing life table... done.