Main survival analysis for taxon profiles

Run the survival analysis per sex and per birth type

Sur_main(
  data.core,
  BirthType = "All",
  models = "GO",
  shape = "simple",
  outlLev1 = NA,
  mindate = "1980-01-01",
  minNsur = 50,
  minlx = 0.1,
  MinBirthKnown = 0.3,
  xMax = 120,
  niter = 25000,
  burnin = 5001,
  thinning = 20,
  nchain = 3,
  ncpus = 2,
  PlotDir = NULL,
  plotname = ""
)

Arguments

data.core

data.frame including at least the following columns anonID, Birth.Date (date), Depart.Date (date), Entry.Date (date), Max.Birth.Date (date), Min.Birth.Date (date), Entry.Type, Depart.Type Sex, birthType, and AnonInstitutionID

BirthType

character When separated the analysis by sex category, the birth type to be selected: Captive, Wild, or All Default = "All"

models

vector of characters names of the basta models to run: "G0", "EX", "LO" and/or "WE". see ?basta for more information. Default = "GO"

shape

character shape of the basta model: "simple", "Makeham", "bathtub". see ?basta for more information. Default = "simple"

outlLev1

numeric Start threshold used to selected for the data: 100%, 99.9, 99 or 95%

mindate

character 'YYYY-MM-DD' Earlier date to include data

minNsur

numeric Minimum number of individual records needed to run the survival analysis. Default = 50

minlx

numeric between 0 and 1. Minimum reached survivorship from the raw Kaplan Meier analysis needed to run the survival analysis. Default = 0.1

MinBirthKnown

numeric between 0 and 1. Minimum proportion of individuals with a known birth month in the data. Default = 0.3

xMax

numeric Maximum age in years Default = 120

niter

numeric. number of MCMC iterations to run. see ?basta for more information. Default = 25000

burnin

numeric Number of iterations removed so that the model has time to converge. see ?basta for more information. Default = 5001

thinning

numeric Number of iteration to run before saving a set of parameters. see ?basta for more information. Default = 20

nchain

numeric Number of chains to run. Default = 5001

ncpus

numeric Number of computer core to use. Default = 2

PlotDir

character Directory to save the plots. Default = NULL, no plot is saved

plotname

character Name used to save the plot. Default = ""

Value

The output of a list including per sex categories:

a summary of the data used:

NGlobal: Number of captive born individuals selected from global collections

NBasta: Number of data (individuals) selected for the BaSTA/survival analysis

Ndead:Number of individuals with known age of death used in the BaSTA/survival analysis

maxAge: Maximum age of known age individuals

maxAlive:Maximum number of years spent ex situ

lxMin:Minimum survivorship reached with the raw Kaplan-Meier model

OutLev: threshold selected for the distribution of time spent alive: 100%, 99.9, 99 or 95%

a logical indicated if the growth analysis was performed

If the survival analysis was not performed, an error and its number (Nerr) are returned: The possibility for this functions are: 1/No raw data and 2/lxMin > minlx 3/NBasta = 0 4/ %known births < MinBirthKnown 5/Data from 1 Institution 6/Nbasta < minNsur, 7/no DIC from Basta.

the basta fit of the best model

the DIC table comparing the different fit of the models

the estimated remaining life expectancy per age If PlotDir is filled, 2 plots are produced: one showing the outliers removed from the data, and one showing the fit of the model on the data.

Examples

data(core)
out <- Sur_main(core, BirthType = "All",
                 models = "GO", shape = "simple",
                 niter = 1000, burnin = 101, thinning = 10, nchain = 3, ncpus = 3)
#> [1] "GO"
#> 
#> Running sequence to find jump SDs... Done
#> Total jump SDs computing time: 9.39 secs.
#> 
#> Multiple simulations started...
#> 
#> snowfall 1.84-6.3 initialized (using snow 0.4-4): parallel execution on 3 CPUs.
#> Library BaSTA2.0 loaded.
#> Library BaSTA2.0 loaded in cluster.
#> 
#> Stopping cluster
#> Simulations finished.
#> Total MCMC computing time: 6.78 secs.
#> 
#> Calculating summary statistics... Done.
#> All parameters converged properly.
#> Constructing life table... done.