Summarize taxon profiles analyses

Produce summary tables and plots of the demographic analyses made for all species

make_summary(
  AnalysisDir,
  SaveDir,
  namefile = "",
  taxaList = "Mammalia",
  BySex = list(Mammalia = c("Male", "Female")),
  Sections = c("sur", "rep", "gro")
)

Arguments

AnalysisDir

character directory where to find the .Rdata files

SaveDir

character directory where to save summary plots and tables

namefile

character Suffix to add to the name of files produced if needed. Default = ""

taxaList

vector of character names of the taxa studied. Default= "Mammalia"

BySex

list of the taxa names indicating the sexes analyzed. Default=list(Mammalia = c("Male", "Female"))

Sections

vector of character names of the sections to update in the taxon profile results: "sur", "rep" and/or "gro". Default = c("sur", "rep", "gro")

Value

It saves summary tables and plots for each Sections and a general summary table. It returns the main summary table

Examples

file = system.file("sci_Animal.csv", package = 'ISRverse')
AnalysisDir  = paste0(dirname(file),'\\Rdata')
SaveDir = paste0(tempdir(check = TRUE),'\\temp')
dir.create(SaveDir)

SummTab <- make_summary(AnalysisDir, SaveDir,
               taxaList = "Reptilia",
               BySex = list(Reptilia = c("Male", "Female")) ,
               Sections = c("sur", 'gro')
)
#> 
#>  Reptilia  -------Species:  Testudo hermanni Progress: 100 %
list.files(SaveDir)
#> [1] "Growth_error.pdf"   "SRGs_Analyses.csv"  "SRGs_Growth.csv"   
#> [4] "SRGs_Survival.csv"  "Survival_error.pdf"

unlink(SaveDir, recursive = TRUE)