Taxon profile report

Run the survival, reproduction and growth analysis for the selected species

tx_report(
  Species,
  Taxa,
  Animal,
  Collection,
  PlotDir = NULL,
  Weights = NULL,
  Parents = NULL,
  Move = NULL,
  Contraceptions = NULL,
  DeathInformation = NULL,
  repout = list(),
  Sections = c("sur", "rep", "gro"),
  SexCats = c("Male", "Female"),
  MinN = 50,
  MinDate = "1980-01-01",
  Global = TRUE,
  MinBirthDate = "1900-01-01",
  ExtractDate = NULL,
  UncertBirth = 365,
  UncertDeath = 365,
  UncertDate = 365,
  MinNSur = 50,
  MaxNSur = NULL,
  MinInstitution = 2,
  MinLx = 0.1,
  MinBirthKnown = 0.3,
  MaxOutl = 100,
  MaxAge = 120,
  MinMLE = 0.1,
  MaxLE = 2,
  ModelsSur = "GO",
  Shape = "bathtub",
  niter = 25000,
  burnin = 5001,
  thinning = 20,
  nchain = 3,
  ncpus = 2,
  MinAge = 0,
  LastDead = FALSE,
  RepSect = c("agemat", "litter"),
  BirthType = "Captive",
  ParentPercDam = 80,
  ParentPercSire = 80,
  MinNLitter = 20,
  NDay = 7,
  MinNRepro = 100,
  MinNPaRepro = 30,
  MinNSeas = 50,
  MinNGro = 100,
  MinNIGro = 50,
  MeasureType = "Live weight",
  ModelsGro = "vonBertalanffy"
)

Arguments

Species

character name of the species studied

Taxa

character name of the taxa studied

Animal

data.frame core data

Collection

data.frame Collection data

PlotDir

character Directory to save the plots. Default: no plot is saved

Weights

data.frame Weights data if the growth analysis is run

Parents

data.frame Parents data if the reproduction analysis is run

Move

data frame Move data including at least the following columns: AnimalAnonID, To, Date

Contraceptions

data.frame Contraception data if the reproduction analysis is run

DeathInformation

data.frame including at least the following columns AnimalAnonID and RelevantDeathInformationType

repout

list Previous result of the taxon profile for this species if it needs to be updated only

Sections

vector of character names of the sections to update in the taxon profile results: "sur", "rep" and/or "gro".

SexCats

character Male, Female or All.

MinN

numeric Minimum number of individuals needed to run the analyses.

MinDate

character 'YYYY-MM-DD' Earliest date to include data

Global

logical Whether only individuals belonging to global Collections should be used.

MinBirthDate

character: Earliest possible date: date used when minimum birth date unknown

ExtractDate

character 'YYYY-MM-DD' Date of data extraction

UncertBirth

numeric: Maximum uncertainty accepted for birth dates, in days

UncertDeath

numeric: Maximum uncertainty accepted for death dates, in days.

UncertDate

numeric: Maximum uncertainty accepted for measurement dates: weight, in days.

MinNSur

numeric Minimum number of individual records needed to run the survival analysis.

MaxNSur

numeric Maximum number of individual records to run the survival analysis.

MinInstitution

numeric Minimum number of institutions that should hold records to run the analyses.

MinLx

numeric Value used for longevity threshold and for checks. between 0 and 1. Minimum reached survivorship from the raw Kaplan Meier analysis. This number avoids running survival analysis if there are too few dead individuals in the data. Lower is better.

MinBirthKnown

numeric between 0 and 1. Minimum proportion of individuals with a known birth month in the data to run the survival analysis.

MaxOutl

numeric Start threshold to select individuals based on the longevity distribution: 100%, 99.9, 99 or 95%. This number must decrease when many errors are expected in longevity data.

MaxAge

numeric Maximum possible age in years. Only used for model predictions. This argument is not used to select data.

MinMLE

numeric Value used for checks. Minimum survivorship allowed at mean life expectancy. Between 0 and 1.

MaxLE

numeric Value used for checks. Maximum remaining life expectancy at last observed age. In years.

ModelsSur

vector of characters Names of the survival basta models to run: "G0", "EX", "LO" and/or "WE". see ?basta for more information.

Shape

character Shape of the survival basta model to run: "simple", "Makeham", "bathtub". see ?basta for more information.

niter

numeric. number of MCMC iterations to run the survival model. see ?basta for more information.

burnin

numeric Number of iterations removed so that the survival model has time to converge. see ?basta for more information.

thinning

numeric Number of iterations to run the survival model before saving a set of parameters. see ?basta for more information.

nchain

numeric Number of chains to run the survival model.

ncpus

numeric Number of computer cores to use.

MinAge

numeric Ages at which the survival analysis should start, in years. see ?basta for more information.

LastDead

logical Whether the longest lived individuals should be considered dead.

RepSect

character Names of the reproductive analyses to run: "agemat", "litter" and/or ...

BirthType

character Captive, Wild, or All.

ParentPercDam

numeric Minimum percentage of parentage probability to include Dam.

ParentPercSire

numeric Minimum percentage of parentage probability to include Sire.

MinNLitter

numeric Minimum number of litters to run the analysis. The data frame for litter size will be produced in all cases.

NDay

numeric Number of consecutive days over which the birth dates of a litter/clutch can be spread.

MinNRepro

numeric Minimum number of birth records needed to run reproductive analyses.

MinNPaRepro

numeric Minimum number of unique parent records needed to run reproductive analyses.

MinNSeas

numeric XXXXXXXXXXX

MinNGro

numeric Minimum number of weights needed to fit the growth models.

MinNIGro

numeric Minimum number of unique individuals needed to fit the growth models.

MeasureType

vector of characters Name of the types of measurement that should be included. Default = all measurement type are included.

ModelsGro

vector of characters Names of the growth models that need to be fit. The following models are supported : logistic, gompertz, chapmanRichards, vonBertalanffy, polynomial.

Value

a list including the results and summary of each analysis. See each function ?Sur_main, ?Rep_main and ?Gro_main for more details about survival, reproduction and growth analyses, respectively.

Examples

file = system.file("sci_Animal.csv", package = 'ISRverse')
ZIMSDirtest = dirname(file)
data <- Load_Zimsdata  (Taxa = "Reptilia",
                       Species = list(Reptilia = "All"),
                       ZIMSDir = ZIMSDirtest,
                       Animal = TRUE, tables = c("Collection","DeathInformation"))
#> ReptiliaLoading Collection of Reptilia.Loading DeathInformation of Reptilia.Reptilia Done.

Animal<- Prep_Animal(data[["Reptilia"]]$Animal, ExtractDate="2024/08/29"  )
PlotDir = paste0(tempdir(check = TRUE),'/temp')
dir.create(PlotDir)

out <- tx_report(Species = "Testudo hermanni", Taxa = "Reptilia",
                 Animal, data$Reptilia$Collection, 
                 BirthType = "All", UncertBirth = 3500,UncertDeath = 3500, 
                 DeathInformation =data$Reptilia$DeathInformation,
                 PlotDir = PlotDir,Sections = c('sur'),
                 SexCats = c("Male", "Female"),
                 ModelsSur = "GO", Shape = "simple",
                 ModelsGro = "vonBertalanffy",
                 niter = 1000, burnin = 101, thinning = 10, nchain = 3, ncpus = 3
)
#>  ****************************  Male  ****************************
#> Survival Running ------------------------------------------------
#>  ****************************  Female  ****************************
#> Survival Running ------------------------------------------------

list.files(PlotDir)
#> [1] "Long_dist" "Survival" 
unlink(PlotDir, recursive = TRUE)