Run Taxon Profiles

Load data needed and run or update demographic analyses for the list of species selected

run_txprofile(
  taxa,
  Species_list,
  ZIMSdir,
  AnalysisDir,
  PlotDir,
  extractDate,
  minDate = "1980-01-01",
  Sections,
  erase_previous = FALSE,
  sexCats = c("Male", "Female"),
  inparallel = FALSE,
  ipara = 1,
  npara = 1,
  minN = 50,
  maxOutl = 99,
  spOutLev = NULL,
  minNsur = 50,
  minlx = 0.1,
  MinBirthKnown = 0.3,
  models_sur = "GO",
  shape = "bathub",
  niter = 25000,
  burnin = 5001,
  thinning = 20,
  nchain = 3,
  ncpus = 2,
  parentProb = 80,
  minNrepro = 100,
  minNparepro = 30,
  minNseas = 50,
  minNgro = 100,
  minNIgro = 50,
  MeasureType = "",
  models_gro = "vonBertalanffy"
)

Arguments

taxa

character the name of the taxa studied

Species_list

vector of character Species selected

ZIMSdir

character directory where to find data

AnalysisDir

character directory where to save results

PlotDir

character Directory to save the plots. Default = NULL, no plot is saved

extractDate

character 'YYYY-MM-DD' Date of data extraction

minDate

character 'YYYY-MM-DD' Earlier date to include data

Sections

vector of character names of the sections to update in the taxon profile results: "sur", "rep" and/or "gro". Default = c("sur", "rep", "gro")

erase_previous

logical whether the current result file should be deleted (before being replaced). Default = FALSE

sexCats

character Male, Female or All Default = "All"

inparallel

logical Whether this function is run in parallel on different computer. In other words: should the species list be divided? Default = FALSE

ipara

numeric Id number of the computer used to select the species to run

npara

numeric number of computers used in parallel

minN

numeric Minimum number of individuals. Default = 50

maxOutl

numeric Maximum threshold for the longevity distribution. Default = NULL

spOutLev

vector of character List of species for which the Maximum threshold for the longevity distribution.XXXXXXXXXXXXXXXXXXXXXXXXX

minNsur

numeric Minimum number of individuals to run the survival analysis. Default = 50

minlx

numeric between 0 and 1. Minimum reached survivorship from the raw Kaplan Meier analysis needed to run the survival analysis. Default = 0.1

MinBirthKnown

numeric between 0 and 1. Minimum proportion of individuals with a known birth month in the data. Default = 0.3

models_sur

vector of characters names of the survival basta models to run: "G0", "EX", "LO" and/or "WE". see ?basta for more information. Default = "GO"

shape

character shape of the survival basta model to run: "simple", "Makeham", "bathtub". see ?basta for more information. Default = "simple"

niter

numeric. number of MCMC iterations to run the survival model. see ?basta for more information. Default = 25000

burnin

numeric Number of iterations removed so that the survival model has time to converge. see ?basta for more information. Default = 5001

thinning

numeric Number of iteration to run the survival model before saving a set of parameters. see ?basta for more information. Default = 20

nchain

numeric Number of chains to run the survival model. Default = 5001

ncpus

numeric Number of computer core to use. Default = 2

parentProb

numeric XXXXXXXXXXX

minNrepro

numeric XXXXXXXXXXX

minNparepro

numeric XXXXXXXXXXX

minNseas

numeric XXXXXXXXXXX

minNgro

numeric Minimum number of weight needed to fit the growth models

minNIgro

numeric Minimum number of unique individuals needed to fit the growth models

MeasureType

vector of characters Name of the type of measurements that should be included. Default = NULL, all measurement type are included.

models_gro

vector of characters indicating the growth models that need to be fit.The following models are supported : logistic, gompertz, chapmanRichards, vonBertalanffy, polynomial. default = "vonBertalanffy"

Value

Save and replace the result file for each specie sin the list. The file is saved in analysisDir\taxa\species.Rdata

Examples

# # Here is an example but use directly your own ZIMSdir(directory to find data) and analysisDir(directory to save analysis)
# file = system.file("coretest.csv", package = 'ISRverse')
# ZIMSdir = dirname(file)
# AnalysisDir = paste0(tempdir(check = TRUE),'\temp')
# PlotDir = paste0(tempdir(check = TRUE),'\temp\plot')
# dir.create(AnalysisDir)
# dir.create(PlotDir)
# 
# #This code run the survival analysis for gorilla
# out <- run_txprofile(taxa = "Mammalia", Species_list = c('Gorilla gorilla'), 
#                           ZIMSdir = RastDir, AnalysisDir = AnalysisDir,
#                           PlotDir = PlotDir,
#                           extractDate = "", 
#                           minDate = "1980-01-01",
#                           Sections = "sur", 
#                           sexCats = c('Male', 'Female'),
#                          niter = 1000, burnin = 101, thinning = 10, nchain = 3, ncpus = 3
# )
# 
# list.files(PlotDir)
# list.files(AnalysisDir)
# 
# unlink(AnalysisDir, recursive = TRUE)

if (FALSE) { # \dontrun{
# Here is an example but use directly your own ZIMSdir(directory to find data) and analysisDir(directory #' to save analysis)
file = system.file("coretest.csv", package = 'ISRverse')
ZIMSdir = dirname(file)
analysisDir = paste0(tempdir(check = TRUE),'\\temp')
PlotDir = paste0(tempdir(check = TRUE),'\\temp\\plot')
dir.create(analysisDir)
dir.create(PlotDir)

#Here I want to split my analysis on two computers
#Run on computer 1:
out <- run_txprofile(taxa = "Mammalia", Species_list = c('Gorilla gorilla', "Capreolus capreolus"), 
                          ZIMSdir = RastDir, analysisDir = analysisDir,
                         PlotDir = PlotDir,
                       extractDate = "", 
                          minDate = "1980-01-01",
                          Sections = "sur", 
                          sexCats = c('Male', 'Female'), 
                          inparallel = TRUE, 
                          ipara = 1, npara = 2
)

#Run on computer 2:
out <- run_txprofile(taxa = "Mammalia", Species_list = c('Gorilla gorilla', "Capreolus capreolus"), 
                          ZIMSdir = RastDir, analysisDir = analysisDir,
                          PlotDir = PlotDir,
                      extractDate = "", 
                          minDate = "1980-01-01",
                          Sections = "sur", 
                          sexCats = c('Male', 'Female'), 
                          inparallel = TRUE,
                           ipara = 2, npara = 2
)
list.files(PlotDir)
list.files(AnalysisDir)

unlink(analysisDir, recursive = TRUE)
} # }