Run Taxon Profiles

Load data needed and run or update demographic analyses for the list of species selected

run_txprofile(
  taxa,
  Species_list,
  ZIMSdir,
  AnalysisDir,
  PlotDir,
  extractDate,
  minDate = "1980-01-01",
  Sections,
  erase_previous = FALSE,
  sexCats = c("Male", "Female"),
  minBirthDate = "1900-01-01",
  inparallel = FALSE,
  ipara = 1,
  npara = 1,
  minN = 50,
  maxOutl = 99,
  spOutLev = NULL,
  Global = TRUE,
  uncert_birth = 365,
  uncert_death = 365,
  uncert_date = 365,
  minNsur = 50,
  maxNsur = NULL,
  minInstitution = 2,
  minlx = 0.1,
  MinBirthKnown = 0.3,
  XMAX = 120,
  Min_MLE = 0.1,
  MaxLE = 2,
  minAge = 0,
  firstyear = FALSE,
  models_sur = "GO",
  shape = "bathtub",
  niter = 25000,
  burnin = 5001,
  thinning = 20,
  nchain = 3,
  ncpus = 2,
  Repsect = c("agemat", "litter"),
  Birth_Type = "Captive",
  lastdead = FALSE,
  parentProb_Sire = 80,
  parentProb_Dam = 80,
  minNlitter = 20,
  Nday = 7,
  minNrepro = 100,
  minNparepro = 30,
  minNseas = 50,
  minNgro = 100,
  minNIgro = 50,
  MeasureType = "Live weight",
  models_gro = "vonBertalanffy"
)

Arguments

taxa

character the name of the taxa studied

Species_list

vector of character Species selected

ZIMSdir

character directory where to find data

AnalysisDir

character directory where to save results

PlotDir

character Directory to save the plots. Default = NULL, no plot is saved

extractDate

character 'YYYY-MM-DD' Date of data extraction

minDate

character 'YYYY-MM-DD' Earlier date to include data

Sections

vector of character names of the sections to update in the taxon profile results: "sur", "rep" and/or "gro". Default = c("sur", "rep", "gro")

erase_previous

logical whether the current result file should be deleted (before being replaced). Default = FALSE

sexCats

character Male, Female or All Default = "All"

minBirthDate

character: Earlier possible date: date used when minimum birth or death date unknown

inparallel

logical Whether this function is run in parallel on different computer. In other words: should the species list be divided? Default = FALSE

ipara

numeric Id number of the computer used to select the species to run

npara

numeric number of computers used in parallel

minN

numeric Minimum number of individuals. Default = 50

maxOutl

numeric Maximum threshold for the longevity distribution. Default = NULL

spOutLev

vector of character List of species for which the Maximum threshold for the longevity distribution.XXXXXXXXXXXXXXXXXXXXXXXXX

Global

logical Whether only individuals belonging to global collections should be used.

uncert_birth

numeric: Maximum uncertainty accepted for birth dates, in days

uncert_death

numeric: Maximum uncertainty accepted for death dates, in days. Default = 365

uncert_date

numeric: Maximum uncertainty accepted for measurement dates: weight, in days. Default = 365

minNsur

numeric Minimum number of individuals to run the survival analysis. Default = 50

maxNsur

numeric Maximum number of individual records to run the survival analysis. Default = NULL

minInstitution

numeric Minimum number of institutions that should hold records to run the analyses. Default = 2

minlx

numeric between 0 and 1. Minimum reached survivorship from the raw Kaplan Meier analysis needed to run the survival analysis. Default = 1

MinBirthKnown

numeric between 0 and 1. Minimum proportion of individuals with a known birth month in the data. Default = 0.3

XMAX

numeric Maximum possible age. Default = 120

Min_MLE

numeric Goodness of fit: Minimum survivorship at Mean life expectancy

MaxLE

numeric Goodness of fit: Maximum remaining life expectancy at max age

minAge

numeric Ages at which the analyses should start. see ?basta for more information. Default = 0

firstyear

logical Whether to do the analysis only on first year. Default = FALSE

models_sur

vector of characters names of the survival basta models to run: "G0", "EX", "LO" and/or "WE". see ?basta for more information. Default = "GO"

shape

character shape of the survival basta model to run: "simple", "Makeham", "bathtub". see ?basta for more information. Default = "simple"

niter

numeric. number of MCMC iterations to run the survival model. see ?basta for more information. Default = 25000

burnin

numeric Number of iterations removed so that the survival model has time to converge. see ?basta for more information. Default = 5001

thinning

numeric Number of iteration to run the survival model before saving a set of parameters. see ?basta for more information. Default = 20

nchain

numeric Number of chains to run the survival model. Default = 5001

ncpus

numeric Number of computer core to use. Default = 2

Repsect

character names of the reproductive analyses to run: "agemat", "litter" and/or ...

Birth_Type

character Captive, Wild, or All. Default = "Captive"

lastdead

logical Whether the longest lived individuals should be considered dead. Default = FALSE

parentProb_Sire

numeric Minimum percentage of parentage probability to include for Sire. Default = 80

parentProb_Dam

numeric Minimum percentage of parentage probability to include for Dam. Default = 80

minNlitter

numeric Minimum number of litters to run the analysis. The data frame for litter size will be produced in all cases. Default = 30

Nday

numeric Number of consecutive days over which the birth dates of a litter/clutch can be spread. Default = 7

minNrepro

numeric Minimum number of birth records needed to run reproductive analyses. Default = 50

minNparepro

numeric Minimum number of unique parent records needed to run reproductive analyses. Default = 30

minNseas

numeric XXXXXXXXXXX

minNgro

numeric Minimum number of weight needed to fit the growth models

minNIgro

numeric Minimum number of unique individuals needed to fit the growth models

MeasureType

vector of characters Name of the type of measurements that should be included. Default = NULL, all measurement type are included.

models_gro

vector of characters indicating the growth models that need to be fit.The following models are supported : logistic, gompertz, chapmanRichards, vonBertalanffy, polynomial. default = "vonBertalanffy"

Value

Save and replace the result file for each specie sin the list. The file is saved in analysisDir\Rdata\taxa_species.Rdata

Examples

# Here is an example but use directly your own ZIMSdir (directory to find data) 
# and analysisDir (directory to save analysis)
file = system.file("sci_Animal.csv", package = 'ISRverse')
ZIMSdir = dirname(file)
AnalysisDir = paste0(tempdir(check = TRUE),'/temp')
PlotDir = paste0(AnalysisDir,'/plot')
# dir.create(AnalysisDir)
dir.create(paste0(AnalysisDir,'/Rdata'), recursive = TRUE)
dir.create(PlotDir)

#This code run the survival analysis
run_txprofile(taxa = "Reptilia", Species_list = "All",
              ZIMSdir = ZIMSdir, AnalysisDir = AnalysisDir,
              PlotDir = PlotDir,
              extractDate = "",
              minDate = "1980-01-01",
              Sections = c("sur"),
              sexCats = c('Male', 'Female'),
              niter = 1000, burnin = 101, thinning = 10, nchain = 3, ncpus = 3
)
#> Loading Data
#> 
#> ====================
#> Class:    Reptilia 
#> Species:  Testudo hermanni 
#> Progress: 100 %
#> =====================
#> Species running...
#>   ****************************  Male  ****************************
#> Survival Running ------------------------------------------------
#> [1] "GO"
#> 
#> Running sequence to find jump SDs... Done
#> Total jump SDs computing time: 16.69 secs.
#> 
#> Multiple simulations started...
#> 
#> R Version:  R version 4.3.2 (2023-10-31 ucrt) 
#> 
#> snowfall 1.84-6.3 initialized (using snow 0.4-4): parallel execution on 3 CPUs.
#> Library BaSTA loaded.
#> Library BaSTA loaded in cluster.
#> Warning: package 'BaSTA' was built under R version 4.3.3
#> 
#> Stopping cluster
#> Simulations finished.
#> Total MCMC computing time: 8.89 secs.
#> 
#> Calculating summary statistics... Done.
#> All parameters converged properly.
#> Constructing life table... done.
#> snowfall 1.84-6.3 initialized (using snow 0.4-4): parallel execution on 3 CPUs.
#> 
#> Stopping cluster
#> snowfall 1.84-6.3 initialized (using snow 0.4-4): parallel execution on 3 CPUs.
#> 
#> Stopping cluster
#> snowfall 1.84-6.3 initialized (using snow 0.4-4): parallel execution on 3 CPUs.
#> 
#> Stopping cluster
#> snowfall 1.84-6.3 initialized (using snow 0.4-4): parallel execution on 3 CPUs.
#> 
#> Stopping cluster
#> snowfall 1.84-6.3 initialized (using snow 0.4-4): parallel execution on 3 CPUs.
#> 
#> Stopping cluster
#> Joining with `by = join_by(Ages)`
#>  ****************************  Female  ****************************
#> Survival Running ------------------------------------------------
#> [1] "GO"
#> 
#> Running sequence to find jump SDs... Done
#> Total jump SDs computing time: 15.74 secs.
#> 
#> Multiple simulations started...
#> 
#> snowfall 1.84-6.3 initialized (using snow 0.4-4): parallel execution on 3 CPUs.
#> Library BaSTA loaded.
#> Library BaSTA loaded in cluster.
#> 
#> Stopping cluster
#> Simulations finished.
#> Total MCMC computing time: 6.35 secs.
#> 
#> Calculating summary statistics... Done.
#> Convergence not reached for some parameters.
#> Constructing life table... done.
#> [1] "more chains"
#> [1] "GO"
#> 
#> Running sequence to find jump SDs... Done
#> Total jump SDs computing time: 15.36 secs.
#> 
#> Multiple simulations started...
#> 
#> Simulations finished.
#> Total MCMC computing time: 24.41 secs.
#> 
#> Calculating summary statistics... Done.
#> All parameters converged properly.
#> Constructing life table... done.
#> snowfall 1.84-6.3 initialized (using snow 0.4-4): parallel execution on 3 CPUs.
#> 
#> Stopping cluster
#> snowfall 1.84-6.3 initialized (using snow 0.4-4): parallel execution on 3 CPUs.
#> 
#> Stopping cluster
#> snowfall 1.84-6.3 initialized (using snow 0.4-4): parallel execution on 3 CPUs.
#> 
#> Stopping cluster
#> snowfall 1.84-6.3 initialized (using snow 0.4-4): parallel execution on 3 CPUs.
#> 
#> Stopping cluster
#> snowfall 1.84-6.3 initialized (using snow 0.4-4): parallel execution on 3 CPUs.
#> 
#> Stopping cluster
#> Joining with `by = join_by(Ages)`
#>  done.

list.files(PlotDir)
#> [1] "Long_dist" "Survival" 
list.files(paste0(AnalysisDir,'\\Rdata'))
#> [1] "Reptilia_Testudo_hermanni.RData"

unlink(AnalysisDir, recursive = TRUE)
unlink(PlotDir, recursive = TRUE)