Run Taxon Profiles

Load data needed and run or update demographic analyses for the list of species selected

run_txprofile(
  Taxa,
  SpeciesList,
  ZIMSDir,
  AnalysisDir,
  PlotDir,
  ExtractDate,
  MinDate = "1980-01-01",
  Sections,
  ErasePrevious = FALSE,
  SexCats = c("Male", "Female"),
  MinBirthDate = "1900-01-01",
  InParallel = FALSE,
  ipara = 1,
  npara = 1,
  MinN = 50,
  MaxOutl = 99,
  spOutLev = NULL,
  Global = TRUE,
  UncertBirth = 365,
  UncertDeath = 365,
  UncertDate = 365,
  MinNSur = 50,
  MaxNSur = NULL,
  MinInstitution = 2,
  MinLx = 0.1,
  MinBirthKnown = 0.3,
  MaxAge = 120,
  MinMLE = 0.1,
  MaxLE = 2,
  MinAge = 0,
  ModelsSur = "GO",
  Shape = "bathtub",
  niter = 25000,
  burnin = 5001,
  thinning = 20,
  nchain = 3,
  ncpus = 2,
  RepSect = c("agemat", "litter"),
  BirthType = "Captive",
  LastDead = FALSE,
  ParentPercSire = 80,
  ParentPercDam = 80,
  MinNLitter = 20,
  NDay = 7,
  MinNRepro = 100,
  MinNPaRepro = 30,
  MinNSeas = 50,
  MinNGro = 100,
  MinNIGro = 50,
  MeasureType = "Live weight",
  ModelsGro = "vonBertalanffy"
)

Arguments

Taxa

character Name of the taxa studied

SpeciesList

vector of character Species selected

ZIMSDir

character Directory where to find data

AnalysisDir

character Directory where to save results

PlotDir

character Directory to save the plots. Default: no plot is saved

ExtractDate

character 'YYYY-MM-DD' Date of data extraction

MinDate

character 'YYYY-MM-DD' Earliest date to include data

Sections

vector of character Names of the sections to update in the results: "sur", "rep" and/or "gro".

ErasePrevious

logical Whether the current result file should be deleted (before being replaced).

SexCats

character Male, Female or All.

MinBirthDate

character: Earliest possible date: date used when minimum birth or death dates are unknown

InParallel

logical Whether this function is run in parallel on different computers. In other words: should the species list be divided?

ipara

numeric ID number of the computer used

npara

numeric Number of computers used in parallel

MinN

numeric Minimum number of individuals needed to run the analysis.

MaxOutl

numeric Start threshold to select individuals based on the longevity distribution: 100%, 99.9, 99 or 95%. This number must decrease when many errors are expected in longevity data.

spOutLev

vector of character List of species for which the Maximum threshold for the longevity distribution is 99.9% instead of MaxOutl.

Global

logical Whether only individuals belonging to global collections should be used.

UncertBirth

numeric: Maximum uncertainty accepted for birth dates, in days

UncertDeath

numeric: Maximum uncertainty accepted for death dates, in days.

UncertDate

numeric: Maximum uncertainty accepted for measurement dates: weight, in days.

MinNSur

numeric Minimum number of individual records to run the survival analysis.

MaxNSur

numeric Maximum number of individual records to run the survival analysis.

MinInstitution

numeric Minimum number of institutions that should hold records to run the analyses.

MinLx

numeric Value used for longevity threshold and for checks. between 0 and 1. Minimum reached survivorship from the raw Kaplan Meier analysis. This number avoids running survival analysis if there are too few dead individuals in the data. Lower is better.

MinBirthKnown

numeric between 0 and 1. Minimum proportion of individuals with a known birth month in the data.

MaxAge

numeric Maximum possible age in years. Only used for model predictions. This argument is not used to select data.

MinMLE

numeric Value used for checks. Minimum survivorship allowed at mean life expectancy. Between 0 and 1.

MaxLE

numeric Value used for checks. Maximum remaining life expectancy at last observed age. In years.

MinAge

numeric Ages at which the analyses should start. see ?basta for more information.

ModelsSur

vector of characters names of the survival basta models to run: "G0", "EX", "LO" and/or "WE". see ?basta for more information.

Shape

character Shape of the survival basta model to run: "simple", "Makeham", "bathtub". see ?basta for more information.

niter

numeric. Number of MCMC iterations to run the survival model. see ?basta for more information.

burnin

numeric Number of iterations removed so that the survival model has time to converge. see ?basta for more information.

thinning

numeric Number of iteration to run the survival model before saving a set of parameters. see ?basta for more information.

nchain

numeric Number of chains to run the survival model.

ncpus

numeric Number of computer core to use.

RepSect

character Names of the reproductive analyses to run: "agemat", "litter" and/or ...

BirthType

character Captive, Wild, or All.

LastDead

logical Whether the longest lived individual should be considered dead.

ParentPercSire

numeric Minimum percentage of parentage probability to include Sire.

ParentPercDam

numeric Minimum percentage of parentage probability to include Dam.

MinNLitter

numeric Minimum number of litters to run the analysis. The data frame for litter size will be produced in all cases.

NDay

numeric Number of consecutive days over which the birth dates of a litter/clutch can be spread

MinNRepro

numeric Minimum number of birth records needed to run reproductive analyses.

MinNPaRepro

numeric Minimum number of unique parent records needed to run reproductive analyses.

MinNSeas

numeric XXXXXXXXXXX

MinNGro

numeric Minimum number of weight needed to fit the growth models

MinNIGro

numeric Minimum number of unique individuals needed to fit the growth models

MeasureType

vector of characters Names of the types of measurement that should be included. Default: all measurement type are included.

ModelsGro

vector of characters Names the growth models that need to be fit.The following models are supported : logistic, gompertz, chapmanRichards, vonBertalanffy, polynomial. default = "vonBertalanffy"

Value

Save and replace the result file for each species in the list. The file is saved in AnalysisDir/Rdata/Taxa_species.Rdata

Examples

# Here is an example but use directly your own ZIMSDir (directory to find data) 
# and AnalysisDir (directory to save the analysis)
file = system.file("sci_Animal.csv", package = 'ISRverse')
ZIMSDir = dirname(file)
AnalysisDir = paste0(tempdir(check = TRUE),'/temp')
PlotDir = paste0(AnalysisDir,'/plot')
# dir.create(AnalysisDir)
dir.create(paste0(AnalysisDir,'/Rdata'), recursive = TRUE)
dir.create(PlotDir)

#This code run the survival analysis
run_txprofile(Taxa = "Reptilia", SpeciesList = "All",
              ZIMSDir = ZIMSDir, AnalysisDir = AnalysisDir,
              PlotDir = PlotDir,
              ExtractDate = "",
              MinDate = "1980-01-01",
              Sections = c("sur"),
              SexCats = c('Male', 'Female'),
              niter = 1000, burnin = 101, thinning = 10, nchain = 3, ncpus = 3
)
#> Loading Data
#> 
#> ====================
#> Class:    Reptilia 
#> Species:  Testudo hermanni 
#> Progress: 100 %
#> =====================
#> Species running...
#>   ****************************  Male  ****************************
#> Survival Running ------------------------------------------------
#> Warning: no non-missing arguments to min; returning Inf
#> [1] "GO"
#> 
#> Running sequence to find jump SDs... Done
#> Total jump SDs computing time: 15.10 secs.
#> 
#> Multiple simulations started...
#> 
#> Warning: Unknown option on commandline: --file
#> snowfall 1.84-6.3 initialized (using snow 0.4-4): parallel execution on 3 CPUs.
#> Library BaSTA loaded.
#> Library BaSTA loaded in cluster.
#> 
#> Stopping cluster
#> Simulations finished.
#> Total MCMC computing time: 6.89 secs.
#> 
#> Calculating summary statistics... Done.
#> All parameters converged properly.
#> Constructing life table... done.
#> Warning: Unknown option on commandline: --file
#> snowfall 1.84-6.3 initialized (using snow 0.4-4): parallel execution on 3 CPUs.
#> 
#> Stopping cluster
#> Joining with `by = join_by(Ages)`
#>  ****************************  Female  ****************************
#> Survival Running ------------------------------------------------
#> Warning: no non-missing arguments to min; returning Inf
#> [1] "GO"
#> 
#> Running sequence to find jump SDs... Done
#> Total jump SDs computing time: 14.81 secs.
#> 
#> Multiple simulations started...
#> 
#> Warning: Unknown option on commandline: --file
#> snowfall 1.84-6.3 initialized (using snow 0.4-4): parallel execution on 3 CPUs.
#> Library BaSTA loaded.
#> Library BaSTA loaded in cluster.
#> 
#> Stopping cluster
#> Simulations finished.
#> Total MCMC computing time: 7.04 secs.
#> 
#> Calculating summary statistics... Done.
#> All parameters converged properly.
#> Constructing life table... done.
#> Warning: Unknown option on commandline: --file
#> snowfall 1.84-6.3 initialized (using snow 0.4-4): parallel execution on 3 CPUs.
#> 
#> Stopping cluster
#> Joining with `by = join_by(Ages)`
#>  done.

list.files(PlotDir)
#> [1] "Long_dist" "Survival" 
list.files(paste0(AnalysisDir,'/Rdata'))
#> [1] "Reptilia_Testudo_hermanni.RData"

unlink(AnalysisDir, recursive = TRUE)
unlink(PlotDir, recursive = TRUE)