Taxon profile report

Run the survival, reproduction and growth analysis for the selected species

tx_report(
  species,
  taxa,
  Animal,
  collection,
  PlotDir = NULL,
  weights = NULL,
  parents = NULL,
  move = NULL,
  contraceptions = NULL,
  DeathInformation = NULL,
  repout = list(),
  Sections = c("sur", "rep", "gro"),
  sexCats = c("Male", "Female"),
  minN = 50,
  minDate = "1980-01-01",
  Global = TRUE,
  minBirthDate = "1900-01-01",
  extractDate = NULL,
  uncert_birth = 365,
  uncert_death = 365,
  uncert_date = 365,
  minNsur = 50,
  maxNsur = NULL,
  minInstitution = 2,
  minlx = 0.1,
  MinBirthKnown = 0.3,
  maxOutl = 100,
  XMAX = 120,
  Min_MLE = 0.1,
  MaxLE = 2,
  firstyear = FALSE,
  models_sur = "GO",
  shape = "bathtub",
  niter = 25000,
  burnin = 5001,
  thinning = 20,
  nchain = 3,
  ncpus = 2,
  minAge = 0,
  lastdead = FALSE,
  Repsect = c("agemat", "litter"),
  Birth_Type = "Captive",
  parentProb_Dam = 80,
  parentProb_Sire = 80,
  minNlitter = 20,
  Nday = 7,
  minNrepro = 100,
  minNparepro = 30,
  minNseas = 50,
  minNgro = 100,
  minNIgro = 50,
  MeasureType = "Live weight",
  models_gro = "vonBertalanffy"
)

Arguments

species

character the name of the species studied

taxa

character the name of the taxa studied

Animal

data.frame core data

collection

data.frame collection data

PlotDir

character Directory to save the plots. Default = NULL, no plot is saved

weights

data.frame weights data if the growth analysis is run

parents

data.frame parents data if the reproduction analysis is run

move

data frame Moves data including at least the following columns: AnimalAnonID, To, Date

contraceptions

data.frame contraception data if the reproduction analysis is run

DeathInformation

data.frame including at least the following columns AnimalAnonID and RelevantDeathInformationType

repout

list Previous result of the taxon profile for this species if it needs to be updated only

Sections

vector of character names of the sections to update in the taxon profile results: "sur", "rep" and/or "gro". Default = c("sur", "rep", "gro")

sexCats

character Male, Female or All Default = "All"

minN

numeric Minimum number of individuals. Default = 50

minDate

character 'YYYY-MM-DD' Earlier date to include data

Global

logical Whether only individuals belonging to global collections should be used.

minBirthDate

character: Earlier possible date: date used when minimum birth or death date unknown

extractDate

character 'YYYY-MM-DD' Date of data extraction

uncert_birth

numeric: Maximum uncertainty accepted for birth dates, in days

uncert_death

numeric: Maximum uncertainty accepted for death dates, in days. Default = 365

uncert_date

numeric: Maximum uncertainty accepted for measurement dates: weight, in days. Default = 365

minNsur

numeric Minimum number of individual records needed to run the survival analysis. Default = 50

maxNsur

numeric Maximum number of individual records to run the survival analysis. Default = NULL

minInstitution

numeric Minimum number of institutions that should hold records to run the analyses. Default = 2

minlx

numeric between 0 and 1. Minimum reached survivorship from the raw Kaplan Meier analysis needed to run the survival analysis. Default = 0.1

MinBirthKnown

numeric between 0 and 1. Minimum proportion of individuals with a known birth month in the data. Default = 0.3

maxOutl

numeric Maximum threshold for the longevity distribution. Default = 100

XMAX

numeric Maximum possible age. Default = 120

Min_MLE

numeric Goodness of fit: Minimum survivorship at Mean life expectancy

MaxLE

numeric Goodness of fit: Maximum remaining life expectancy at max age

firstyear

logical Whether to do the analysis only on first year. Default = FALSE

models_sur

vector of characters names of the survival basta models to run: "G0", "EX", "LO" and/or "WE". see ?basta for more information. Default = "GO"

shape

character shape of the survival basta model to run: "simple", "Makeham", "bathtub". see ?basta for more information. Default = "simple"

niter

numeric. number of MCMC iterations to run the survival model. see ?basta for more information. Default = 25000

burnin

numeric Number of iterations removed so that the survival model has time to converge. see ?basta for more information. Default = 5001

thinning

numeric Number of iteration to run the survival model before saving a set of parameters. see ?basta for more information. Default = 20

nchain

numeric Number of chains to run the survival model. Default = 5001

ncpus

numeric Number of computer core to use. Default = 2

minAge

numeric Ages at which the analyses should start. see ?basta for more information. Default = 0

lastdead

logical Whether the longest lived individuals should be considered dead. Default = FALSE

Repsect

character names of the reproductive analyses to run: "agemat", "litter" and/or ...

Birth_Type

character Captive, Wild, or All. Default = "Captive"

parentProb_Dam

numeric Minimum percentage of parentage probability to include for Dam. Default = 80

parentProb_Sire

numeric Minimum percentage of parentage probability to include for Sire. Default = 80

minNlitter

numeric Minimum number of litters to run the analysis. The data frame for litter size will be produced in all cases. Default = 30

Nday

numeric Number of consecutive days over which the birth dates of a litter/clutch can be spread. Default = 7

minNrepro

numeric Minimum number of birth records needed to run reproductive analyses. Default = 50

minNparepro

numeric Minimum number of unique parent records needed to run reproductive analyses. Default = 30

minNseas

numeric XXXXXXXXXXX

minNgro

numeric Minimum number of weight needed to fit the growth models

minNIgro

numeric Minimum number of unique individuals needed to fit the growth models

MeasureType

vector of characters Name of the type of measurements that should be included. Default = NULL, all measurement type are included.

models_gro

vector of characters indicating the growth models that need to be fit.The following models are supported : logistic, gompertz, chapmanRichards, vonBertalanffy, polynomial. default = "vonBertalanffy"

Value

a list including the results and summary of each analysis

Examples

file = system.file("sci_Animal.csv", package = 'ISRverse')
ZIMSdirtest = dirname(file)
data <- Load_Zimsdata  (taxa = "Reptilia",
                       species = list(Reptilia = "All"),
                       ZIMSdir = ZIMSdirtest,
                       Animal = TRUE, tables = c("Collection","DeathInformation"))
#> ReptiliaLoading Collection of Reptilia.Loading DeathInformation of Reptilia.Reptilia Done.

Animal<- Prep_Animal(data[["Reptilia"]]$Animal, extractDate="2024/08/29"  )
PlotDir = paste0(tempdir(check = TRUE),'\\temp')
dir.create(PlotDir)

out <- tx_report(species = "Testudo hermanni", taxa = "Reptilia",
                 Animal, data$Reptilia$Collection, 
                 Birth_Type = "All", uncert_birth = 3500,uncert_death = 3500, 
                 DeathInformation =data$Reptilia$DeathInformation,
                 PlotDir = PlotDir,Sections = c('sur'),
                 sexCats = c("Male", "Female"),
                 models_sur = "GO", shape = "simple",
                 models_gro = "vonBertalanffy",
                 niter = 1000, burnin = 101, thinning = 10, nchain = 3, ncpus = 3
)
#>  ****************************  Male  ****************************
#> Survival Running ------------------------------------------------
#>  ****************************  Female  ****************************
#> Survival Running ------------------------------------------------

list.files(PlotDir)
#> [1] "Long_dist" "Survival" 
unlink(PlotDir, recursive = TRUE)