Taxon profile report

Run the survival, reproduction and growth analysis for the selected species

tx_report(
  species,
  taxa,
  core,
  collection,
  PlotDir = NULL,
  weights = NULL,
  parents = NULL,
  contraceptions = NULL,
  repout = list(),
  Sections = c("sur", "rep", "gro"),
  sexCats = c("Male", "Female"),
  minN = 50,
  minDate = "1980-01-01",
  extractDate = NULL,
  minNsur = 50,
  minlx = 0.1,
  MinBirthKnown = 0.3,
  maxOutl = 100,
  models_sur = "GO",
  shape = "bathtub",
  niter = 25000,
  burnin = 5001,
  thinning = 20,
  nchain = 3,
  ncpus = 2,
  parentProb = 80,
  minNrepro = 100,
  minNparepro = 30,
  minNseas = 50,
  minNgro = 100,
  minNIgro = 50,
  MeasureType = "",
  models_gro = "vonBertalanffy"
)

Arguments

species

character the name of the species studied

taxa

character the name of the taxa studied

core

data.frame core data

collection

data.frame collection data

PlotDir

character Directory to save the plots. Default = NULL, no plot is saved

weights

data.frame weights data if the growth analysis is run

parents

data.frame parents data if the reproduction analysis is run

contraceptions

data.frame contraception data if the reproduction analysis is run

repout

list Previous result of the taxon profile for this species if it needs to be updated only

Sections

vector of character names of the sections to update in the taxon profile results: "sur", "rep" and/or "gro". Default = c("sur", "rep", "gro")

sexCats

character Male, Female or All Default = "All"

minN

numeric Minimum number of individuals. Default = 50

minDate

character 'YYYY-MM-DD' Earlier date to include data

extractDate

character 'YYYY-MM-DD' Date of data extraction

minNsur

numeric Minimum number of individual records needed to run the survival analysis. Default = 50

minlx

numeric between 0 and 1. Minimum reached survivorship from the raw Kaplan Meier analysis needed to run the survival analysis. Default = 0.1

MinBirthKnown

numeric between 0 and 1. Minimum proportion of individuals with a known birth month in the data. Default = 0.3

maxOutl

numeric Maximum threshold for the longevity distribution. Default = 100

models_sur

vector of characters names of the survival basta models to run: "G0", "EX", "LO" and/or "WE". see ?basta for more information. Default = "GO"

shape

character shape of the survival basta model to run: "simple", "Makeham", "bathtub". see ?basta for more information. Default = "simple"

niter

numeric. number of MCMC iterations to run the survival model. see ?basta for more information. Default = 25000

burnin

numeric Number of iterations removed so that the survival model has time to converge. see ?basta for more information. Default = 5001

thinning

numeric Number of iteration to run the survival model before saving a set of parameters. see ?basta for more information. Default = 20

nchain

numeric Number of chains to run the survival model. Default = 5001

ncpus

numeric Number of computer core to use. Default = 2

parentProb

numeric XXXXXXXXXXX

minNrepro

numeric XXXXXXXXXXX

minNparepro

numeric XXXXXXXXXXX

minNseas

numeric XXXXXXXXXXX

minNgro

numeric Minimum number of weight needed to fit the growth models

minNIgro

numeric Minimum number of unique individuals needed to fit the growth models

MeasureType

vector of characters Name of the type of measurements that should be included. Default = NULL, all measurement type are included.

models_gro

vector of characters indicating the growth models that need to be fit.The following models are supported : logistic, gompertz, chapmanRichards, vonBertalanffy, polynomial. default = "vonBertalanffy"

Value

a list including the results and summary of each analysis

Examples

data(core)
data(collection)
PlotDir = paste0(tempdir(check = TRUE),'\\temp')
dir.create(PlotDir)

out <- tx_report(species = "Gorilla gorilla", taxa = "Mammalia",
                 core, collection, 
                 PlotDir = PlotDir,Sections = c('sur'),
                 sexCats = c("Male", "Female"),
                 models_sur = "GO", shape = "simple",
                 models_gro = "vonBertalanffy",
                 niter = 1000, burnin = 101, thinning = 10, nchain = 3, ncpus = 3
)
#> [1] "GO"
#> 
#> Running sequence to find jump SDs... Done
#> Total jump SDs computing time: 3.03 secs.
#> 
#> Multiple simulations started...
#> 
#> Warning: Unknown option on commandline: --file
#> snowfall 1.84-6.3 initialized (using snow 0.4-4): parallel execution on 3 CPUs.
#> Library BaSTA loaded.
#> Library BaSTA loaded in cluster.
#> 
#> Stopping cluster
#> Simulations finished.
#> Total MCMC computing time: 3.92 secs.
#> 
#> Calculating summary statistics... Done.
#> All parameters converged properly.
#> Constructing life table... done.

list.files(PlotDir)
#> character(0)
unlink(PlotDir, recursive = TRUE)